+++ OpenGLCourseApp/Light.h
@@ -9,12 +9,11 @@ class Light
 {
 public:
 	Light();
-	Light(GLfloat shadowWidth, GLfloat shadowHeight,
+	Light(GLuint shadowWidth, GLuint shadowHeight,
 		GLfloat red, GLfloat green, GLfloat blue,
 		GLfloat aIntensity, GLfloat dIntensity);
 	ShadowMap* GetShadowMap() { return shadowMap; }
 	~Light();
-
 protected:
 	glm::vec3 color;
 	GLfloat ambientIntensity;
@@ -22,4 +21,3 @@ protected:
 	glm::mat4 lightProj;
 	ShadowMap* shadowMap;
 };
-
