+++ OpenGLCourseApp/SpotLight.cpp
@@ -1,7 +1,5 @@
 #include "SpotLight.h"
 
-
-
 SpotLight::SpotLight() : PointLight()
 {
 	direction = glm::vec3(0.0f, -1.0f, 0.0f);
@@ -10,7 +8,7 @@ SpotLight::SpotLight() : PointLight()
 	isOn = true;
 }
 
-SpotLight::SpotLight(GLfloat shadowWidth, GLfloat shadowHeight,
+SpotLight::SpotLight(GLuint shadowWidth, GLuint shadowHeight,
 	GLfloat near, GLfloat far,
 	GLfloat red, GLfloat green, GLfloat blue,
 	GLfloat aIntensity, GLfloat dIntensity,
@@ -20,7 +18,6 @@ SpotLight::SpotLight(GLfloat shadowWidth, GLfloat shadowHeight,
 	GLfloat edg) : PointLight(shadowWidth, shadowHeight, near, far, red, green, blue, aIntensity, dIntensity, xPos, yPos, zPos, con, lin, exp)
 {
 	direction = glm::normalize(glm::vec3(xDir, yDir, zDir));
-
 	edge = edg;
 	procEdge = cosf(glm::radians(edge));
 }
@@ -31,22 +28,20 @@ void SpotLight::UseLight(GLuint ambientIntensityLocation, GLuint ambientColorLoc
 	GLuint edgeLocation)
 {
 	glUniform3f(ambientColorLocation, color.x, color.y, color.z);
-
 	if (isOn)
 	{
 		glUniform1f(ambientIntensityLocation, ambientIntensity);
 		glUniform1f(diffuseIntensityLocation, diffuseIntensity);
 	}
-	else {
+	else
+	{
 		glUniform1f(ambientIntensityLocation, 0.0f);
 		glUniform1f(diffuseIntensityLocation, 0.0f);
 	}
-
 	glUniform3f(positionLocation, position.x, position.y, position.z);
 	glUniform1f(constantLocation, constant);
 	glUniform1f(linearLocation, linear);
 	glUniform1f(exponentLocation, exponent);
-
 	glUniform3f(directionLocation, direction.x, direction.y, direction.z);
 	glUniform1f(edgeLocation, procEdge);
 }
